+++ src/BPE.cc
@@ -3,9 +3,11 @@
 #include <algorithm>
 #include <fstream>
 #include <limits>
+#include <list>
 
 #include "onmt/unicode/Unicode.h"
 #include "onmt/CaseModifier.h"
+#include "onmt/Tokenizer.h"
 
 namespace onmt
 {
@@ -299,4 +301,5 @@ namespace onmt
     }
 
   }
+
 }